move list pools to pools file
